Fix race condition in delay mechanism of DoT and DoH.
Because start_wait_timer_for_tcp_finish() is called as soon as data is
written on the socket, _finish_session_timer is started as soon as
the handshake begins. But because no data is send at this point, if the
configured s_delay is short enough, e.g. 1ms, the timer will conclude
that the session has finished and will wrongly close.